requirements-server.txt
flask
flask-socketio
eventlet
numpy
scipy
orjson
pyyaml
